+++ src/main/java/com/ssafy/questory/controller/MemberAuthController.java
@@ -3,18 +3,17 @@ package com.ssafy.questory.controller;
 import com.ssafy.questory.domain.Member;
 import com.ssafy.questory.dto.request.member.MemberModifyPasswordRequestDto;
 import com.ssafy.questory.dto.request.member.MemberModifyRequestDto;
+import com.ssafy.questory.dto.request.member.MemberSearchRequestDto;
+import com.ssafy.questory.dto.response.member.MemberSearchResponseDto;
 import com.ssafy.questory.dto.response.member.auth.MemberInfoResponseDto;
 import com.ssafy.questory.service.MemberAuthService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
 
@@ -58,5 +57,13 @@ public class MemberAuthController {
         return ResponseEntity.ok().body(Map.of(
                 "message", "탈퇴하였습니다."
         ));
+    }
+
+    @GetMapping("/search")
+    @Operation(summary = "유저 검색", description = "이메일을 통해 유저를 검색합니다.")
+    public ResponseEntity<Page<MemberSearchResponseDto>> search(
+            @AuthenticationPrincipal(expression = "member") Member member,
+            @ModelAttribute MemberSearchRequestDto memberSearchRequestDto) {
+        return ResponseEntity.ok(memberAuthService.search(member, memberSearchRequestDto));
     }
 }
